fix uint32_t

+++ firmware/events.cpp
@@ -16,7 +16,7 @@ void bench(uint8_t subsys, uint8_t index) {
   data[2] = subsys;
   data[4] = index;
   // bool ret = can.send(0x77000C, data, sizeof(data), true);
-  bool ret = can.send(bench_test_packet_ids_e::ECU_CAN_BUS_USER_CONTROL, data, sizeof(data), true);
+  bool ret = can.send((uint32_t)bench_test_packet_ids_e::ECU_CAN_BUS_USER_CONTROL, data, sizeof(data), true);
   char sys[8] = { 0 };
   switch (subsys) {
     case 0x12:
@@ -100,7 +100,7 @@ void StartStop(lv_event_t *e) {
   uint8_t data[] = { (uint8_t)bench_test_magic_numbers_e::BENCH_HEADER, 0x00, 0x14, 0x00, 0x09, 0x00 };
   // 0x77000C 0x66 0x00 0x14 0x00 0x09 0x00 start/stop engine
   // bool ret = can.send(0x77000C, data, sizeof(data), true);
-  bool ret = can.send(bench_test_packet_ids_e::ECU_CAN_BUS_USER_CONTROL, data, sizeof(data), true); 
+  bool ret = can.send((uint32_t)bench_test_packet_ids_e::ECU_CAN_BUS_USER_CONTROL, data, sizeof(data), true); 
   lv_label_set_text_fmt(ui_debugStatus, "Engine Start/stop send status: %s", ret ? "OK" : "FAIL");
 }
